ticktock-web/requirements.txt
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
markovify==0.9.4
minio==7.1.11
psycopg2-binary==2.9.3
python-magic
